requirements.txt
beautifulsoup4==4.9.3
Flask==2.2.2
numpy== 1.19.5
pandas== 1.1.5
scikit_learn==1.1.3
googlesearch_python==1.0.1
whois==0.9.13
gunicorn==20.1.0
requests==2.25.1
Flask-Mail
